# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sporteam API",
    "version": "2.0.0",
    "description": "API RESTful para la gestión de jugadores, entrenadores y clubs de fútbol amateur",
    "contact": {
      "name": "Roman Dono Perez"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "Token JWT de autenticación"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN_ROLE",
              "USER_ROLE",
              "JUGADOR_ROLE",
              "ENTRENADOR_ROLE",
              "CLUB_ROLE"
            ]
          },
          "estado": {
            "type": "boolean"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "google": {
            "type": "boolean"
          },
          "estadoDeportivo": {
            "type": "string",
            "nullable": true
          },
          "clubId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateUserInput": {
        "type": "object",
        "required": [
          "nombre",
          "apellidos",
          "email",
          "password",
          "role"
        ],
        "properties": {
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "USER_ROLE",
              "JUGADOR_ROLE",
              "ENTRENADOR_ROLE",
              "CLUB_ROLE"
            ]
          },
          "estadoDeportivo": {
            "type": "string"
          },
          "zona": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array de IDs de zona"
          },
          "club": {
            "type": "string",
            "description": "ID del club"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuarios con paginación",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 9
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "usuarios": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error al recuperar usuarios"
          }
        }
      }
    },
    "/api/usuario/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener usuario por ID",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "usuario": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar un usuario",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "usuario": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar un usuario",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
    "/api/usuario": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "usuario": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          }
        }
      }
    }
  },
...
}